Drop fs import in core lib
