+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@skalenetwork/ima-js",
-  "version": "2.0.0-experimental.0",
+  "version": "2.0.0",
   "description": "Simple TS/JS library to interact with SKALE IMA",
   "main": "build/index.js",
   "types": "build/index.d.ts",
